CLOSED - # 235: Weight Gain per day should be more precise. http://localhost:8000/tracker/ticket/235

+++ tracker/trunk/com.verticon.tracker/src/com/verticon/tracker/Animal.java
@@ -347,7 +347,7 @@ public interface Animal extends EObject {
 	 * @model transient="true" changeable="false" volatile="true" derived="true"
 	 * @generated
 	 */
-	Integer getWeightGainPerDay();
+	Double getWeightGainPerDay();
 
 	/**
 	 * <!-- begin-user-doc -->
