+++ tracker/trunk/com.verticon.tracker/src/com/verticon/tracker/WeighIn.java
@@ -77,7 +77,7 @@ public interface WeighIn extends Event {
 	 * @model transient="true" changeable="false" volatile="true" derived="true"
 	 * @generated
 	 */
-	Integer getWeightGainPerDay();
+	Double getWeightGainPerDay();
 
 	/**
 	 * <!-- begin-user-doc -->
